Authoring API WIP: added license headers and fixed pom.xml
