+++ bb-integration-tests/src/test/java/com/cognifide/qa/bb/aem/api/ScreenplayApiTest.java
@@ -1,4 +1,23 @@
-package com.cognifide.qa.bb.aem.authoring;
+/*-
+ * #%L
+ * Bobcat
+ * %%
+ * Copyright (C) 2018 Cognifide Ltd.
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package com.cognifide.qa.bb.aem.api;
 
 import static com.cognifide.qa.bb.api.syntax.Assertions.in;
 import static com.cognifide.qa.bb.api.syntax.Assertions.seeThat;
@@ -8,12 +27,12 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openqa.selenium.WebElement;
 
-import com.cognifide.qa.bb.aem.api.actors.Author;
-import com.cognifide.qa.bb.aem.authoring.pageobjects.PageHero;
+import com.cognifide.qa.bb.aem.api.pageobjects.PageHero;
 import com.cognifide.qa.bb.api.actions.basic.Navigate;
+import com.cognifide.qa.bb.api.actors.Bobcat;
 import com.cognifide.qa.bb.api.actors.abilities.PerformBasicOperations;
-import com.cognifide.qa.bb.api.states.basic.StateInObject;
 import com.cognifide.qa.bb.api.states.assertions.Visibility;
+import com.cognifide.qa.bb.api.states.basic.StateInObject;
 import com.cognifide.qa.bb.core.TestModule;
 import com.cognifide.qa.bb.junit.Modules;
 import com.cognifide.qa.bb.junit.TestRunner;
@@ -24,7 +43,7 @@ import com.google.inject.Inject;
 public class ScreenplayApiTest {
 
   @Inject
-  private Author author;
+  private Bobcat author;
 
   @Test
   public void apiSyntaxCompiles() {
@@ -45,7 +64,8 @@ public class ScreenplayApiTest {
     author.should(seeThat(Visibility.of(PageHero.class, PageHero::button), is(true)));
 
     // direct WebElement check
-    PageHero object = author.thatHasAnAbilityTo(PerformBasicOperations.class).instantiate(PageHero.class);
+    PageHero object =
+        author.thatCan(PerformBasicOperations.class).instantiate(PageHero.class);
     author.should(seeThat(Visibility.of(object.button()), is(true)));
   }
 }
